fix(types): explicit typing for notify args
The main issue here is that Notify is a function type with multiple overloads, meaning it can accept different sets of parameters. TypeScript resolves function overloads at the call site, not within the function implementation. This makes it difficult to infer a generic tuple type for the parameters in a way that works across all overloads.

Workarounds
- Explicit Typing
- Explicit Overload Handling
- Conditional Type Checking